fix: resolve all clippy warnings across workspace + repo hygiene
- 修复 44 处 clippy 告警（while_let_loop、unnecessary_unwrap、manual_clamp、needless_range_loop、field_reassign_with_default、ptr_arg、missing_safety_doc、single_match、needless_update、module_inception、too_many_arguments 等）

- 为 FFI 插件入口补充 # Safety 文档

- 移除误跟踪的 target2/ 构建产物目录

+++ crates/drafftink-quiz/src/ui/quiz_panel.rs
@@ -603,9 +603,9 @@ impl QuizPanel {
                 .next()
                 .and_then(|c| {
                     if c.is_ascii_uppercase() {
-                        Some((c as u8 - b'A') as u8)
+                        Some(c as u8 - b'A')
                     } else if c.is_ascii_digit() {
-                        Some((c as u8 - b'0') as u8)
+                        Some(c as u8 - b'0')
                     } else {
                         None
                     }
@@ -617,9 +617,9 @@ impl QuizPanel {
                     .chars()
                     .filter_map(|c| {
                         if c.is_ascii_uppercase() {
-                            Some((c as u8 - b'A') as u8)
+                            Some(c as u8 - b'A')
                         } else if c.is_ascii_digit() {
-                            Some((c as u8 - b'0') as u8)
+                            Some(c as u8 - b'0')
                         } else {
                             None
                         }
